+++ android-components/components/browser/engine-gecko/src/main/java/mozilla/components/browser/engine/gecko/GeckoEngineSession.kt
@@ -1071,6 +1071,7 @@ class GeckoEngineSession(
             session: GeckoSession,
             url: String?,
             geckoPermissions: List<ContentPermission>,
+            hasUserGesture: Boolean,
         ) {
             this@GeckoEngineSession.geckoPermissions = geckoPermissions
             if (url == null) {
@@ -1104,7 +1105,7 @@ class GeckoEngineSession(
             }
             // Reset the status of current page being product or not when user navigates away.
             notifyObservers { onProductUrlChange(false) }
-            notifyObservers { onLocationChange(url) }
+            notifyObservers { onLocationChange(url, hasUserGesture) }
         }
 
         override fun onLoadRequest(