+++ android-components/components/browser/engine-gecko/src/test/java/mozilla/components/browser/engine/gecko/GeckoEngineSessionTest.kt
@@ -296,13 +296,17 @@ class GeckoEngineSessionTest {
         val engineSession = GeckoEngineSession(runtime, geckoSessionProvider = geckoSessionProvider)
 
         var observedUrl = ""
+        var observedUserGesture = true
         var observedCanGoBack = false
         var observedCanGoForward = false
         var cookieBanner = CookieBannerHandlingStatus.HANDLED
         var displaysProduct = false
         engineSession.register(
             object : EngineSession.Observer {
-                override fun onLocationChange(url: String) { observedUrl = url }
+                override fun onLocationChange(url: String, hasUserGesture: Boolean) {
+                    observedUrl = url
+                    observedUserGesture = hasUserGesture
+                }
                 override fun onNavigationStateChange(canGoBack: Boolean?, canGoForward: Boolean?) {
                     canGoBack?.let { observedCanGoBack = canGoBack }
                     canGoForward?.let { observedCanGoForward = canGoForward }
@@ -318,8 +322,9 @@ class GeckoEngineSessionTest {
 
         captureDelegates()
 
-        navigationDelegate.value.onLocationChange(mock(), "http://mozilla.org", emptyList())
+        navigationDelegate.value.onLocationChange(mock(), "http://mozilla.org", emptyList(), false)
         assertEquals("http://mozilla.org", observedUrl)
+        assertEquals(false, observedUserGesture)
         assertEquals(CookieBannerHandlingStatus.NO_DETECTED, cookieBanner)
         // TO DO: add a positive test case after a test endpoint is implemented in desktop (Bug 1846341)
         assertEquals(false, displaysProduct)
@@ -856,22 +861,22 @@ class GeckoEngineSessionTest {
         var observedUrl = ""
         engineSession.register(
             object : EngineSession.Observer {
-                override fun onLocationChange(url: String) { observedUrl = url }
+                override fun onLocationChange(url: String, hasUserGesture: Boolean) { observedUrl = url }
             },
         )
 
         captureDelegates()
 
-        navigationDelegate.value.onLocationChange(mock(), "about:blank", emptyList())
+        navigationDelegate.value.onLocationChange(mock(), "about:blank", emptyList(), false)
         assertEquals("", observedUrl)
 
-        navigationDelegate.value.onLocationChange(mock(), "about:blank", emptyList())
+        navigationDelegate.value.onLocationChange(mock(), "about:blank", emptyList(), false)
         assertEquals("", observedUrl)
 
-        navigationDelegate.value.onLocationChange(mock(), "https://www.mozilla.org", emptyList())
+        navigationDelegate.value.onLocationChange(mock(), "https://www.mozilla.org", emptyList(), false)
         assertEquals("https://www.mozilla.org", observedUrl)
 
-        navigationDelegate.value.onLocationChange(mock(), "about:blank", emptyList())
+        navigationDelegate.value.onLocationChange(mock(), "about:blank", emptyList(), false)
         assertEquals("about:blank", observedUrl)
     }
 
@@ -881,7 +886,7 @@ class GeckoEngineSessionTest {
         captureDelegates()
         assertTrue(session.initialLoad)
 
-        navigationDelegate.value.onLocationChange(mock(), "https://mozilla.org", emptyList())
+        navigationDelegate.value.onLocationChange(mock(), "https://mozilla.org", emptyList(), false)
         assertFalse(session.initialLoad)
 
         navigationDelegate.value.onLoadRequest(mock(), mockLoadRequest("moz-extension://1234-test"))
@@ -890,16 +895,16 @@ class GeckoEngineSessionTest {
         var observedUrl = ""
         session.register(
             object : EngineSession.Observer {
-                override fun onLocationChange(url: String) { observedUrl = url }
+                override fun onLocationChange(url: String, hasUserGesture: Boolean) { observedUrl = url }
             },
         )
-        navigationDelegate.value.onLocationChange(mock(), "about:blank", emptyList())
+        navigationDelegate.value.onLocationChange(mock(), "about:blank", emptyList(), false)
         assertEquals("", observedUrl)
 
-        navigationDelegate.value.onLocationChange(mock(), "https://www.mozilla.org", emptyList())
+        navigationDelegate.value.onLocationChange(mock(), "https://www.mozilla.org", emptyList(), false)
         assertEquals("https://www.mozilla.org", observedUrl)
 
-        navigationDelegate.value.onLocationChange(mock(), "about:blank", emptyList())
+        navigationDelegate.value.onLocationChange(mock(), "about:blank", emptyList(), false)
         assertEquals("about:blank", observedUrl)
     }
 
@@ -929,10 +934,10 @@ class GeckoEngineSessionTest {
         geckoResult.complete(true)
 
         assertNull(engineSession.currentUrl)
-        navigationDelegate.value.onLocationChange(geckoSession, "https://www.mozilla.org", emptyList())
+        navigationDelegate.value.onLocationChange(geckoSession, "https://www.mozilla.org", emptyList(), false)
         assertEquals("https://www.mozilla.org", engineSession.currentUrl)
 
-        navigationDelegate.value.onLocationChange(geckoSession, "https://www.firefox.com", emptyList())
+        navigationDelegate.value.onLocationChange(geckoSession, "https://www.firefox.com", emptyList(), false)
         assertEquals("https://www.firefox.com", engineSession.currentUrl)
     }
 
@@ -942,7 +947,7 @@ class GeckoEngineSessionTest {
 
         captureDelegates()
 
-        navigationDelegate.value.onLocationChange(geckoSession, "https://www.mozilla.org", listOf(mock()))
+        navigationDelegate.value.onLocationChange(geckoSession, "https://www.mozilla.org", listOf(mock()), false)
 
         assertTrue(engineSession.geckoPermissions.isNotEmpty())
     }
@@ -953,7 +958,7 @@ class GeckoEngineSessionTest {
 
         captureDelegates()
 
-        navigationDelegate.value.onLocationChange(geckoSession, null, listOf(mock()))
+        navigationDelegate.value.onLocationChange(geckoSession, null, listOf(mock()), false)
 
         assertTrue(engineSession.geckoPermissions.isNotEmpty())
     }
@@ -979,7 +984,7 @@ class GeckoEngineSessionTest {
         verify(historyTrackingDelegate, never()).onTitleChanged(anyString(), anyString())
 
         // This sets the currentUrl.
-        navigationDelegate.value.onLocationChange(geckoSession, "https://www.mozilla.com", emptyList())
+        navigationDelegate.value.onLocationChange(geckoSession, "https://www.mozilla.com", emptyList(), false)
 
         contentDelegate.value.onTitleChange(geckoSession, "Hello World!")
         verify(historyTrackingDelegate).onTitleChanged(eq("https://www.mozilla.com"), eq("Hello World!"))
@@ -1036,7 +1041,7 @@ class GeckoEngineSessionTest {
 
         engineSession.register(
             object : EngineSession.Observer {
-                override fun onLocationChange(url: String) { observedUrl = url }
+                override fun onLocationChange(url: String, hasUserGesture: Boolean) { observedUrl = url }
                 override fun onTitleChange(title: String) { observedTitle = title }
             },
         )
@@ -1070,7 +1075,7 @@ class GeckoEngineSessionTest {
             ),
         )
 
-        navigationDelegate.value.onLocationChange(geckoSession, emptyPageUrl, emptyList())
+        navigationDelegate.value.onLocationChange(geckoSession, emptyPageUrl, emptyList(), false)
         contentDelegate.value.onTitleChange(geckoSession, emptyPageUrl)
 
         historyDelegate.value.onVisited(
@@ -1110,7 +1115,7 @@ class GeckoEngineSessionTest {
         verify(historyTrackingDelegate, never()).onPreviewImageChange(anyString(), anyString())
 
         // This sets the currentUrl.
-        navigationDelegate.value.onLocationChange(geckoSession, "https://www.mozilla.com", emptyList())
+        navigationDelegate.value.onLocationChange(geckoSession, "https://www.mozilla.com", emptyList(), false)
 
         contentDelegate.value.onPreviewImage(geckoSession, previewImageUrl)
         verify(historyTrackingDelegate).onPreviewImageChange(eq("https://www.mozilla.com"), eq(previewImageUrl))