+++ android-components/components/browser/state/src/test/java/mozilla/components/browser/state/engine/EngineObserverTest.kt
@@ -142,7 +142,7 @@ class EngineObserverTest {
             override fun exitFullScreenMode() {}
             override fun purgeHistory() {}
             override fun loadData(data: String, mimeType: String, encoding: String) {
-                notifyObservers { onLocationChange(data) }
+                notifyObservers { onLocationChange(data, false) }
                 notifyObservers { onProgress(100) }
                 notifyObservers { onLoadingStateChange(true) }
                 notifyObservers { onNavigationStateChange(true, true) }
@@ -155,7 +155,7 @@ class EngineObserverTest {
                 flags: LoadUrlFlags,
                 additionalHeaders: Map<String, String>?,
             ) {
-                notifyObservers { onLocationChange(url) }
+                notifyObservers { onLocationChange(url, false) }
                 notifyObservers { onProgress(100) }
                 notifyObservers { onLoadingStateChange(true) }
                 notifyObservers { onNavigationStateChange(true, true) }
@@ -479,7 +479,7 @@ class EngineObserverTest {
 
         assertEquals("Mozilla", store.state.tabs[0].content.title)
 
-        observer.onLocationChange("https://getpocket.com")
+        observer.onLocationChange("https://getpocket.com", false)
         store.waitUntilIdle()
 
         assertEquals("", store.state.tabs[0].content.title)
@@ -506,7 +506,7 @@ class EngineObserverTest {
 
         assertEquals("Mozilla", store.state.tabs[0].content.title)
 
-        observer.onLocationChange("https://www.mozilla.org")
+        observer.onLocationChange("https://www.mozilla.org", false)
         store.waitUntilIdle()
 
         assertEquals("Mozilla", store.state.tabs[0].content.title)
@@ -533,7 +533,7 @@ class EngineObserverTest {
 
         assertEquals("Mozilla", store.state.tabs[0].content.title)
 
-        observer.onLocationChange("https://www.mozilla.org/#something")
+        observer.onLocationChange("https://www.mozilla.org/#something", false)
         store.waitUntilIdle()
 
         assertEquals("Mozilla", store.state.tabs[0].content.title)
@@ -560,7 +560,7 @@ class EngineObserverTest {
 
         assertEquals(previewImageUrl, store.state.tabs[0].content.previewImageUrl)
 
-        observer.onLocationChange("https://getpocket.com")
+        observer.onLocationChange("https://getpocket.com", false)
         store.waitUntilIdle()
 
         assertNull(store.state.tabs[0].content.previewImageUrl)
@@ -588,12 +588,12 @@ class EngineObserverTest {
 
         assertEquals(previewImageUrl, store.state.tabs[0].content.previewImageUrl)
 
-        observer.onLocationChange("https://www.mozilla.org")
+        observer.onLocationChange("https://www.mozilla.org", false)
         store.waitUntilIdle()
 
         assertEquals(previewImageUrl, store.state.tabs[0].content.previewImageUrl)
 
-        observer.onLocationChange("https://www.mozilla.org/#something")
+        observer.onLocationChange("https://www.mozilla.org/#something", false)
         store.waitUntilIdle()
 
         assertEquals(previewImageUrl, store.state.tabs[0].content.previewImageUrl)
@@ -681,7 +681,7 @@ class EngineObserverTest {
 
         assertEquals(manifest, store.state.tabs[0].content.webAppManifest)
 
-        observer.onLocationChange("https://getpocket.com")
+        observer.onLocationChange("https://getpocket.com", false)
         store.waitUntilIdle()
 
         assertNull(store.state.tabs[0].content.webAppManifest)
@@ -709,7 +709,7 @@ class EngineObserverTest {
 
         assertEquals(listOf(request), store.state.tabs[0].content.permissionRequestsList)
 
-        observer.onLocationChange("https://getpocket.com")
+        observer.onLocationChange("https://getpocket.com", false)
         store.waitUntilIdle()
 
         assertEquals(emptyList<PermissionRequest>(), store.state.tabs[0].content.permissionRequestsList)
@@ -737,7 +737,7 @@ class EngineObserverTest {
 
         assertEquals(listOf(request), store.state.tabs[0].content.permissionRequestsList)
 
-        observer.onLocationChange("https://www.mozilla.org/hello.html")
+        observer.onLocationChange("https://www.mozilla.org/hello.html", false)
         store.waitUntilIdle()
 
         assertEquals(listOf(request), store.state.tabs[0].content.permissionRequestsList)
@@ -765,7 +765,7 @@ class EngineObserverTest {
 
         assertEquals(manifest, store.state.tabs[0].content.webAppManifest)
 
-        observer.onLocationChange("https://www.mozilla.org/hello.html")
+        observer.onLocationChange("https://www.mozilla.org/hello.html", false)
         store.waitUntilIdle()
 
         assertEquals(manifest, store.state.tabs[0].content.webAppManifest)
@@ -797,12 +797,12 @@ class EngineObserverTest {
 
         assertEquals(manifest, store.state.tabs[0].content.webAppManifest)
 
-        observer.onLocationChange("https://www.mozilla.org/hello/page2.html")
+        observer.onLocationChange("https://www.mozilla.org/hello/page2.html", false)
         store.waitUntilIdle()
 
         assertEquals(manifest, store.state.tabs[0].content.webAppManifest)
 
-        observer.onLocationChange("https://www.mozilla.org/hello.html")
+        observer.onLocationChange("https://www.mozilla.org/hello.html", false)
         store.waitUntilIdle()
         assertNull(store.state.tabs[0].content.webAppManifest)
     }
@@ -1512,7 +1512,7 @@ class EngineObserverTest {
         )
 
         val observer = EngineObserver("test-id", store)
-        observer.onLocationChange("https://www.mozilla.org/en-US/")
+        observer.onLocationChange("https://www.mozilla.org/en-US/", false)
 
         store.waitUntilIdle()
 
@@ -1697,11 +1697,58 @@ class EngineObserverTest {
         observer.onLoadUrl()
         store.waitUntilIdle()
 
-        middleware.assertNotDispatched(ContentAction.UpdateSearchTermsAction::class)
         middleware.assertLastAction(ContentAction.UpdateIsSearchAction::class) { action ->
             assertEquals(false, action.isSearch)
             assertEquals("test-id", action.sessionId)
         }
+    }
+
+    @Test
+    fun `GIVEN a search is performed WHEN the location is changed by user interaction THEN the search terms are cleared`() {
+        val middleware = CaptureActionsMiddleware<BrowserState, BrowserAction>()
+        val store = BrowserStore(
+            initialState = BrowserState(
+                tabs = listOf(
+                    createTab("https://www.mozilla.org", id = "test-id"),
+                ),
+            ),
+            middleware = listOf(middleware),
+        )
+
+        store.dispatch(ContentAction.UpdateIsSearchAction("test-id", true))
+        store.waitUntilIdle()
+
+        val observer = EngineObserver("test-id", store)
+        observer.onLocationChange("testUrl", true)
+        store.waitUntilIdle()
+
+        middleware.assertLastAction(ContentAction.UpdateSearchTermsAction::class) { action ->
+            assertEquals("", action.searchTerms)
+            assertEquals("test-id", action.sessionId)
+        }
+    }
+
+
+    @Test
+    fun `GIVEN a search is performed WHEN the location is changed without user interaction THEN the search terms are not cleared`() {
+        val middleware = CaptureActionsMiddleware<BrowserState, BrowserAction>()
+        val store = BrowserStore(
+            initialState = BrowserState(
+                tabs = listOf(
+                    createTab("https://www.mozilla.org", id = "test-id"),
+                ),
+            ),
+            middleware = listOf(middleware),
+        )
+
+        store.dispatch(ContentAction.UpdateIsSearchAction("test-id", true))
+        store.waitUntilIdle()
+
+        val observer = EngineObserver("test-id", store)
+        observer.onLocationChange("testUrl", false)
+        store.waitUntilIdle()
+
+        middleware.assertNotDispatched(ContentAction.UpdateSearchTermsAction::class)
     }
 
     @Test